changing package name to logic and organize the files
