+++ bin/.gitignore
@@ -3,3 +3,4 @@
 /common/
 /gui/
 /logic/
+/donotenterdrinksorfood/
