+++ src/logic/Supplier.java
@@ -1,4 +1,4 @@
-package donotenterdrinksorfood;
+package logic;
 
 public class Supplier {
 	private int ID;
